[COMMON] scsi: ufs: fix clang warning
This patch resolve below waring.

warning: implicit conversion from 'unsigned long' to 'int' changes
         value from 18446744073709551615 to -1 [-Wconstant-conversion]

+++ drivers/scsi/ufs/ufshcd.c
@@ -2718,7 +2718,7 @@ static bool ufshcd_get_dev_cmd_tag(struct ufs_hba *hba, int *tag_out)
 	do {
 		tmp = ~hba->lrb_in_use;
 		tmp &= BITMAP_LAST_WORD_MASK(hba->nutrs);
-		tag = (tmp) ? __fls(tmp) : ~0ul;
+		tag = (int)(tmp) ? __fls(tmp) : ~0;
 		if (tag >= hba->nutrs)
 			goto out;
 	} while (test_and_set_bit_lock(tag, &hba->lrb_in_use));
